+++ hs-abci-sdk/package.yaml
@@ -55,7 +55,6 @@ dependencies:
 - cereal
 - cereal-text
 - containers
-- conduit
 - cryptonite
 - data-default-class
 - errors
@@ -98,35 +97,33 @@ library:
   - -Wredundant-constraints
   exposed-modules:
   - Tendermint.SDK.Application
-  - Tendermint.SDK.Aeson
+  - Tendermint.SDK.Application.App
+  - Tendermint.SDK.Application.Module
+  - Tendermint.SDK.Application.Handlers
+  - Tendermint.SDK.BaseApp.BaseApp
+  - Tendermint.SDK.BaseApp.Errors
+  - Tendermint.SDK.BaseApp.Events
+  - Tendermint.SDK.BaseApp.Gas
+  - Tendermint.SDK.BaseApp.Query
+  - Tendermint.SDK.BaseApp.Query.Class
+  - Tendermint.SDK.BaseApp.Query.Delayed
+  - Tendermint.SDK.BaseApp.Query.Router
+  - Tendermint.SDK.BaseApp.Query.Store
+  - Tendermint.SDK.BaseApp.Query.Types
+  - Tendermint.SDK.BaseApp.Query.Client
+  - Tendermint.SDK.BaseApp.Logger
+  - Tendermint.SDK.BaseApp.Logger.Katip
+  - Tendermint.SDK.BaseApp.Store
+  - Tendermint.SDK.BaseApp.Store.AuthTreeStore
+  - Tendermint.SDK.BaseApp.Store.RawStore
+  - Tendermint.SDK.BaseApp.Store.Scope
   - Tendermint.SDK.Auth
-  - Tendermint.SDK.BaseApp
   - Tendermint.SDK.Codec
   - Tendermint.SDK.Crypto
-  - Tendermint.SDK.Errors
-  - Tendermint.SDK.Events
-  - Tendermint.SDK.Query
-  - Tendermint.SDK.Query.Class
-  - Tendermint.SDK.Query.Delayed
-  - Tendermint.SDK.Query.Router
-  - Tendermint.SDK.Query.Store
-  - Tendermint.SDK.Query.Types
-  - Tendermint.SDK.Query.Client
-  - Tendermint.SDK.Logger
-  - Tendermint.SDK.Logger.Katip
-  - Tendermint.SDK.Store
-  - Tendermint.SDK.Store.AuthTreeStore
-  - Tendermint.SDK.Store.RawStore
-  - Tendermint.SDK.Store.Scope
-  - Tendermint.SDK.Subscription
   - Tendermint.SDK.Types.Message
   - Tendermint.SDK.Types.Transaction
   - Tendermint.SDK.Types.Address
   - Tendermint.SDK.Types.TxResult
-  - Tendermint.SDK.TxRouter
-  - Tendermint.SDK.Module
-  - Tendermint.SDK.Gas
-  - Tendermint.SDK.Handlers
 
   generated-exposed-modules:
   - Proto.Types.Transaction
