--- hs-abci-sdk/src/Tendermint/SDK/Aeson.hs
@@ -1,9 +1,0 @@
-module Tendermint.SDK.Aeson
-  ( defaultSDKAesonOptions
-  ) where
-
-import           Data.Aeson        (Options)
-import           Data.Aeson.Casing (aesonDrop, snakeCase)
-
-defaultSDKAesonOptions :: String -> Options
-defaultSDKAesonOptions prefix = aesonDrop (length prefix) snakeCase
