+++ hs-abci-sdk/src/Tendermint/SDK/Auth.hs
@@ -2,23 +2,23 @@
 
 module Tendermint.SDK.Auth where
 
-import           Data.Bifunctor               (first)
+import           Data.Bifunctor                (first)
 import           Data.Proxy
-import qualified Data.Serialize               as Serialize
-import           Data.Serialize.Text          ()
-import           Data.String.Conversions      (cs)
-import           Data.Text                    (Text)
+import qualified Data.Serialize                as Serialize
+import           Data.Serialize.Text           ()
+import           Data.String.Conversions       (cs)
+import           Data.Text                     (Text)
 import           Data.Word
-import           GHC.Generics                 (Generic)
-import           GHC.TypeLits                 (symbolVal)
+import           GHC.Generics                  (Generic)
+import           GHC.TypeLits                  (symbolVal)
 import           Polysemy
-import           Polysemy.Error               (Error, mapError)
-import           Tendermint.SDK.Codec         (HasCodec (..))
-import           Tendermint.SDK.Errors        (AppError (..), IsAppError (..))
-import           Tendermint.SDK.Query         (Queryable (..))
-import           Tendermint.SDK.Store         (IsKey (..), RawStore,
-                                               StoreKey (..), get, put)
-import           Tendermint.SDK.Types.Address (Address)
+import           Polysemy.Error                (Error, mapError)
+import           Tendermint.SDK.BaseApp.Errors (AppError (..), IsAppError (..))
+import           Tendermint.SDK.BaseApp.Query  (Queryable (..))
+import           Tendermint.SDK.BaseApp.Store  (IsKey (..), RawStore,
+                                                StoreKey (..), get, put)
+import           Tendermint.SDK.Codec          (HasCodec (..))
+import           Tendermint.SDK.Types.Address  (Address)
 
 --------------------------------------------------------------------------------
