+++ hs-abci-sdk/src/Tendermint/SDK/BaseApp/Store.hs
@@ -1,0 +1,7 @@
+module Tendermint.SDK.BaseApp.Store
+  ( module Tendermint.SDK.BaseApp.Store.RawStore
+  , module Tendermint.SDK.BaseApp.Store.Scope
+  ) where
+
+import           Tendermint.SDK.BaseApp.Store.RawStore
+import           Tendermint.SDK.BaseApp.Store.Scope
