--- hs-abci-sdk/src/Tendermint/SDK/Store.hs
@@ -1,7 +1,0 @@
-module Tendermint.SDK.Store
-  ( module Tendermint.SDK.Store.RawStore
-  , module Tendermint.SDK.Store.Scope
-  ) where
-
-import           Tendermint.SDK.Store.RawStore
-import           Tendermint.SDK.Store.Scope
