+++ hs-abci-sdk/src/Tendermint/SDK/Types/Address.hs
@@ -1,13 +1,13 @@
 module Tendermint.SDK.Types.Address where
 
-import           Control.Lens             (iso)
-import qualified Crypto.Secp256k1         as Crypto
-import qualified Data.Aeson               as A
-import qualified Data.ByteArray.HexString as Hex
-import           Data.ByteString          (ByteString)
-import           GHC.Generics             (Generic)
-import           Tendermint.SDK.Query     (FromQueryData)
-import           Tendermint.SDK.Store     (RawKey (..))
+import           Control.Lens                 (iso)
+import qualified Crypto.Secp256k1             as Crypto
+import qualified Data.Aeson                   as A
+import qualified Data.ByteArray.HexString     as Hex
+import           Data.ByteString              (ByteString)
+import           GHC.Generics                 (Generic)
+import           Tendermint.SDK.BaseApp.Query (FromQueryData)
+import           Tendermint.SDK.BaseApp.Store (RawKey (..))
 
 -- | Used as a unique identifier for an account.
 newtype Address =
